+++ All.ui/components/removeBG/removeBG.js
@@ -11,9 +11,16 @@ Component({
    * 组件的初始数据
    */
   data: {
-
+    isRBG: false, // 是否开始消除背景
   },
 
+  attached() {
+    setTimeout(() => {
+      this.setData({
+        isRBG: true
+      })
+    }, 2000)
+  },
   /**
    * 组件的方法列表
    */